create __main__ module for cluster fixture
